feat: make it possible to make a guaranteed subscription to a user creation event
There was a problem that if for some reason
the service was not available or something with rabbitMQ,
the user was not saved in the database,
which would then affect the whole application.
It was decided to create a separate queue for user-service
and link it to exchange from auth-service.
We also added a confirmation of message receipt in RabbitMQ
and resending the message in case of failure,
which also gives a guarantee that the user will be saved in the database.

!!! Note:
For correct operation, you must manually configure queues
and connect them to exchange in the RabbitMQ interface